requirements.txt
crewai>=0.86.0
fastapi>=0.115.0
uvicorn>=0.34.0
line-bot-sdk>=3.14.0
python-dotenv>=1.0.0
streamlit>=1.40.0
